% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_to_amino}
\alias{codon_to_amino}
\title{Standard Codon Code}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_to_amino
}
\description{
Standard translation table from codons to amino acids
}
\keyword{datasets}
