% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_rna.R
\name{convert_to_rna}
\alias{convert_to_rna}
\title{Convert DNA sequence to RNA}
\usage{
convert_to_rna(dna_seq)
}
\arguments{
\item{DNA_seq}{A character string representing a DNA sequence}
}
\value{
Outputs an RNA sequence, where each T of the input sequence is swapped with U
}
\description{
Convert DNA sequence to RNA
}
\examples{
convert_to_RNA("AGAGGTTTTTTCGA") # Expected output: "AGAGGUUUUUUCGA"
}
