% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Dna_to_rna
Give a dna sequence}
\usage{
dna_to_rna(sequence)
}
\arguments{
\item{sequence}{Get the rna sequence}
}
\value{
rna
}
\description{
Dna_to_rna
Give a dna sequence
}
