% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DnaToRna.R
\name{DnaToRna}
\alias{DnaToRna}
\title{Convert DNA to RNA}
\usage{
DnaToRna(dnaSequence)
}
\arguments{
\item{dnaSequence}{}
}
\value{
rnaSequence
}
\description{
Convert DNA to RNA
}
\examples{
dnaSequence(ATATATA)
AUAUAUA

}
