% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_codons.R
\name{find_codons}
\alias{find_codons}
\title{Extract each individual codons from a sequence.}
\usage{
find_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A character string representing an RNA sequence}

\item{start}{The starting position from which the codon extraction will start.}
}
\value{
A character vector containing codons
}
\description{
Extract each individual codons from a sequence.
}
\examples{
find_codons("UUUUCUUAUAGGUGA",1) # Expected output: c("UUU","UCU","UAU","AGG","UGA")
}
