% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genCode}
\alias{genCode}
\title{genetic code}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
genCode
}
\description{
genetic code (description)
}
\keyword{datasets}
