% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_sequence.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{DNA generation}
\usage{
generate_dna_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{create a dna sequencew of a specified length}
}
\value{
random_dna_sequence
returns a DNA sequence composed of a random sequence of nuclotides
}
\description{
DNA generation
}
