% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodons.R
\name{getCodons}
\alias{getCodons}
\title{Get the codons of a RNA sequence}
\usage{
getCodons(rnaSequence, start = 1)
}
\arguments{
\item{start}{}
}
\value{
codons
}
\description{
Get the codons of a RNA sequence
}
\examples{
getCodons("AUGGACUGAUCGAUGU", start = 1)
[1] "AUG" "GAC" "UGA" "UCG" "AUG"

}
