% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_plot.R
\name{peptide_plot}
\alias{peptide_plot}
\title{Plot amino acid composition of peptide sequence}
\usage{
peptide_plot(peptide_seq)
}
\arguments{
\item{peptide_seq}{A character string representing a peptide sequence.}
}
\value{
A ggplot object representing a bar plot of the counts of each amino acid in the provided peptide sequence.
}
\description{
Plot amino acid composition of peptide sequence
}
