% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFrequencies.R
\name{plotFrequencies}
\alias{plotFrequencies}
\title{Plot amino acid frequencies}
\usage{
plotFrequencies(aminoAcids)
}
\arguments{
\item{aaSequence}{Amino acid sequence as string}
}
\description{
Plot amino acid frequencies
}
