% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDNASequence.R
\name{randomDNASequence}
\alias{randomDNASequence}
\title{Generate random DNA sequence}
\usage{
randomDNASequence(length)
}
\arguments{
\item{length}{Length of the sequence}
}
\value{
Random DNA sequence (string)
}
\description{
Generate random DNA sequence
}
