% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate codons into protein sequence}
\usage{
translate(codons)
}
\arguments{
\item{codons}{A character vector containing codons}
}
\value{
A character string representing the translated protein sequence
}
\description{
Translate codons into protein sequence
}
\examples{
translate(c("UUU", "UCU")) # Expected output: "FS"
}
