% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateRNASequence.R
\name{translateRNASequence}
\alias{translateRNASequence}
\title{Translate RNA into protein}
\usage{
translateRNASequence(codons)
}
\arguments{
\item{codons}{Vector of codons}
}
\value{
Amino acid sequence (string)
}
\description{
Translate RNA into protein
}
