% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_RNA.R
\name{convert_to_RNA}
\alias{convert_to_RNA}
\title{Convert DNA sequence to RNA}
\usage{
convert_to_RNA(DNA_seq)
}
\arguments{
\item{DNA_seq}{A DNA sequence}
}
\value{
Outputs an RNA sequence, where each T is swapped with U
}
\description{
Convert DNA sequence to RNA
}
