% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_codons.R
\name{find_codons}
\alias{find_codons}
\title{Extract each individual codons from a sequence.}
\usage{
find_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{a RNA sequence}

\item{start}{where in the sequence you want to start}
}
\value{
a string of codons
}
\description{
Extract each individual codons from a sequence.
}
\examples{
find_codons("UUUUCUUAUAGGUGA",1)
 output= c("UUU","UCU","UAU","AGG","UGA"))
}
