% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna_sequence.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{generate_dna_sequence}
\usage{
generate_dna_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{}
}
\value{
Generate a random sequence of "A", "T", "G", "C" with the specified length
}
\description{
generate_dna_sequence
}
\examples{
generate_dna_sequence(2)

}
