% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_five.R
\name{peptide_plot}
\alias{peptide_plot}
\title{Plot aminoacid composition of peptide sequence}
\usage{
peptide_plot(peptide_seq)
}
\arguments{
\item{peptide_seq}{A peptide sequence}
}
\value{
Plot of counts of each aminoacid in the sequence.
}
\description{
Plot aminoacid composition of peptide sequence
}
