% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate codons into protein sequence}
\usage{
translate(codons)
}
\arguments{
\item{codons}{A vector of codons}
}
\value{
the translated protein sequence in string format
}
\description{
Translate codons into protein sequence
}
